requirements.txt
face_recognition==1.3.0
faiss_cpu==1.10.0
fastapi==0.115.14
numpy==2.3.1
pydantic==2.11.7
